requirements.txt
pytest == 8.2.2
pycryptodome == 3.20.0
gmpy2 == 2.1.5
ecdsa == 0.19.0
numpy == 1.26.4
libnum == 1.7.1
